adding of larger than uint8_t value in available gets truncated

+++ common_arm/usb_cdc.c
@@ -716,7 +716,7 @@ uint32_t usb_read(uint8_t *data, size_t len) {
 
 static uint8_t usb_read_ng_buffer[64] = {0};
 static uint8_t usb_read_ng_bufoffset = 0;
-static uint8_t usb_read_ng_buflen = 0;
+static uint16_t usb_read_ng_buflen = 0;
 
 bool usb_read_ng_has_buffered_data(void) {
     return usb_read_ng_buflen > 0;
